feat: updated README, added more SSO functionality
