+++ src/core/server/app/middleware/passport/index.ts
@@ -59,14 +59,16 @@ export function createPassport(
 }
 
 interface LogoutToken {
-  jti: string;
-  exp: number;
+  jti?: string;
+  exp?: number;
 }
 
-const LogoutTokenSchema = Joi.object().keys({
-  jti: Joi.string(),
-  exp: Joi.number(),
-});
+const LogoutTokenSchema = Joi.object()
+  .keys({
+    jti: Joi.string().default(undefined),
+    exp: Joi.number().default(undefined),
+  })
+  .optionalKeys(["jti", "exp"]);
 
 export async function handleLogout(redis: Redis, req: Request, res: Response) {
   // Extract the token from the request.
@@ -88,13 +90,14 @@ export async function handleLogout(redis: Redis, req: Request, res: Response) {
 
   // Grab the JTI from the decoded token.
   const { jti, exp }: LogoutToken = validate(LogoutTokenSchema, decoded);
-
-  // Compute the number of seconds that the token will be valid for.
-  const validFor = exp - now.valueOf() / 1000;
-  if (validFor > 0) {
-    // Invalidate the token, the expiry is in the future and it needs to be
-    // revoked.
-    await revokeJWT(redis, jti, validFor, now);
+  if (jti && exp) {
+    // Compute the number of seconds that the token will be valid for.
+    const validFor = exp - now.valueOf() / 1000;
+    if (validFor > 0) {
+      // Invalidate the token, the expiry is in the future and it needs to be
+      // revoked.
+      await revokeJWT(redis, jti, validFor, now);
+    }
   }
 
   // Clear the cookie.
