+++ src/core/server/app/middleware/passport/strategies/verifiers/sso.spec.ts
@@ -52,7 +52,6 @@ describe("SSOUserProfileSchema", () => {
       email: "email",
       username: "username",
       avatar: "avatar",
-      displayName: "displayName",
     };
 
     expect(validate(SSOUserProfileSchema, profile)).toEqual(profile);
@@ -64,19 +63,6 @@ describe("SSOUserProfileSchema", () => {
       id: "id",
       email: "email",
       username: "username",
-      displayName: "displayName",
-    };
-
-    expect(validate(SSOUserProfileSchema, profile)).toEqual(profile);
-    expect(isSSOToken({ user: profile })).toEqual(true);
-  });
-
-  it("allows an empty displayName", () => {
-    const profile = {
-      id: "id",
-      email: "email",
-      username: "username",
-      avatar: "avatar",
     };
 
     expect(validate(SSOUserProfileSchema, profile)).toEqual(profile);
