+++ src/core/server/app/middleware/passport/strategies/verifiers/sso.ts
@@ -1,5 +1,5 @@
 import Joi from "joi";
-import { isNil } from "lodash";
+import { isNil, isUndefined } from "lodash";
 import { Db } from "mongodb";
 
 import { validate } from "coral-server/app/request/body";
@@ -9,13 +9,24 @@ import {
   GQLUSER_ROLE,
 } from "coral-server/graph/tenant/schema/__generated__/types";
 import { Tenant } from "coral-server/models/tenant";
-import { retrieveUserWithProfile, SSOProfile } from "coral-server/models/user";
+import {
+  retrieveUserWithProfile,
+  SSOProfile,
+  updateUserFromSSO,
+} from "coral-server/models/user";
 import { insert } from "coral-server/services/users";
 
 import {
+  getSSOProfile,
+  needsSSOUpdate,
+} from "coral-server/models/user/helpers";
+import {
+  isJWTRevoked,
   SymmetricSigningAlgorithm,
   verifyJWT,
 } from "coral-server/services/jwt";
+import { AugmentedRedis } from "coral-server/services/redis";
+import { DateTime } from "luxon";
 import { Verifier } from "../jwt";
 
 export interface SSOStrategyOptions {
@@ -26,29 +37,40 @@ export interface SSOUserProfile {
   id: string;
   email: string;
   username: string;
-  avatar?: string;
 }
 
 export interface SSOToken {
+  jti?: string;
+  exp?: number;
+  iat?: number;
   user: SSOUserProfile;
 }
 
-export const SSOUserProfileSchema = Joi.object()
-  .keys({
-    id: Joi.string().required(),
-    email: Joi.string().required(),
-    username: Joi.string().required(),
-    avatar: Joi.string().default(undefined),
-    displayName: Joi.string().default(undefined),
-  })
-  .optionalKeys(["avatar", "displayName"]);
+export function isSSOToken(token: SSOToken | object): token is SSOToken {
+  const { error } = Joi.validate(token, SSOTokenSchema);
+  return isNil(error);
+}
 
-export const SSOTokenSchema = Joi.object().keys({
-  user: SSOUserProfileSchema.required(),
+export const SSOUserProfileSchema = Joi.object().keys({
+  id: Joi.string().required(),
+  email: Joi.string()
+    .lowercase()
+    .required(),
+  username: Joi.string().required(),
 });
+
+export const SSOTokenSchema = Joi.object()
+  .keys({
+    jti: Joi.string().default(undefined),
+    exp: Joi.number().default(undefined),
+    iat: Joi.number().default(undefined),
+    user: SSOUserProfileSchema.required(),
+  })
+  .optionalKeys(["jti", "exp", "iat"]);
 
 export async function findOrCreateSSOUser(
   mongo: Db,
+  redis: AugmentedRedis,
   tenant: Tenant,
   integration: GQLSSOAuthIntegration,
   token: SSOToken,
@@ -59,19 +81,31 @@ export async function findOrCreateSSOUser(
     throw new Error("token is malformed, missing user claim");
   }
 
-  // Unpack/validate the token content.
-  const { id, email, username, avatar }: SSOUserProfile = validate(
-    SSOUserProfileSchema,
-    token.user
-  );
+  // Validate the token content.
+  const decodedToken: SSOToken = validate(SSOTokenSchema, token);
 
-  const profile: SSOProfile = {
-    type: "sso",
-    id,
-  };
+  // Unpack the token.
+  const {
+    jti,
+    exp,
+    user: { id, email, username },
+    iat,
+  } = decodedToken;
+
+  // If the token has a JTI and EXP claim, then it can be logged out. Check to
+  // see if it was revoked.
+  if (jti && exp && (await isJWTRevoked(redis, jti))) {
+    return null;
+  }
+
+  // Compute the last issued at time stamp.
+  const lastIssuedAt = iat ? DateTime.fromSeconds(iat).toJSDate() : now;
 
   // Try to lookup user given their id provided in the `sub` claim.
-  let user = await retrieveUserWithProfile(mongo, tenant.id, profile);
+  let user = await retrieveUserWithProfile(mongo, tenant.id, {
+    type: "sso",
+    id,
+  });
   if (!user) {
     if (!integration.allowRegistration) {
       // Registration is disabled, so we can't create the user user here.
@@ -79,6 +113,12 @@ export async function findOrCreateSSOUser(
     }
 
     // FIXME: (wyattjoh) implement rules! Not all users should be able to create an account via this method.
+
+    const profile: SSOProfile = {
+      type: "sso",
+      id,
+      lastIssuedAt,
+    };
 
     // Create the new user, as one didn't exist before!
     user = await insert(
@@ -88,32 +128,42 @@ export async function findOrCreateSSOUser(
         username,
         role: GQLUSER_ROLE.COMMENTER,
         email,
-        avatar,
         profiles: [profile],
       },
       now
     );
+  } else if (iat && needsSSOUpdate(decodedToken.user, user)) {
+    // Get the SSO Profile.
+    const profile = getSSOProfile(user);
+    if (profile && profile.lastIssuedAt < lastIssuedAt) {
+      // The token presented to us has a newer issue date than the one
+      // associated with this profile, we should update the user with new
+      // details.
+      user = await updateUserFromSSO(
+        mongo,
+        tenant.id,
+        user.id,
+        { email, username },
+        lastIssuedAt
+      );
+    }
   }
-
-  // TODO: (wyattjoh) possibly update the user profile if the remaining details mismatch?
 
   return user;
 }
 
-export function isSSOToken(token: SSOToken | object): token is SSOToken {
-  const { error } = Joi.validate(token, SSOTokenSchema);
-  return isNil(error);
-}
-
 export interface SSOVerifierOptions {
   mongo: Db;
+  redis: AugmentedRedis;
 }
 
 export class SSOVerifier implements Verifier<SSOToken> {
   private mongo: Db;
+  private redis: AugmentedRedis;
 
-  constructor({ mongo }: SSOVerifierOptions) {
+  constructor({ mongo, redis }: SSOVerifierOptions) {
     this.mongo = mongo;
+    this.redis = redis;
   }
 
   public supports(token: SSOToken | object, tenant: Tenant): token is SSOToken {
@@ -147,6 +197,13 @@ export class SSOVerifier implements Verifier<SSOToken> {
       now
     );
 
-    return findOrCreateSSOUser(this.mongo, tenant, integration, token, now);
+    return findOrCreateSSOUser(
+      this.mongo,
+      this.redis,
+      tenant,
+      integration,
+      token,
+      now
+    );
   }
 }
